requirements.txt
django==2.0.1
djangorestframework==3.7.7
djangorestframework-gis==0.12
django-filter==1.1.0
psycopg2==2.7.3.2